Reorganizing package directories
